Treat empty string glob as Value::Nothing
Original goal: Prevent commands like `rm ""` from taking the `""` to mean the
current working directory.

Analysis: The issue is that when evaluated, the empty string argument gets
concatenated with the current workign directory. It's not even a globbing
issue, as this happens before the argument is passed through globbing.

Proposed resolution: Treat empty glob patterns as an empty value when
evaluated.